All the elements of the whiteboard itself have been finished according to the design Vitus and Jacob drew up. Now, the WhiteBoard is composed of WhiteBoardColumns. There are 5 different subclasses of WhiteBoardColumn. These classes are "UserStoryColumn", "ToDoColumn", "InProgressColumn", "InVerificationColumn", and "CompleteColumn". Each column has an "OnDropBehavior" that states what needs to happen if a Notecard/Postit is dropped into it.
Also, I have reworked the UserStoryServiceImpl so that now, each UserStory actually owns its Tasks. This is a dependent relationship (meaning that if you delete a UserStory, you delete all of its Tasks. This has required quite a bit of research and we will need to test this class extensively! Sorry Hameed for running over some of your work on Task in order to do this.