---Mass Testing files update--- Complete: LoginInfo, NoteCard, PostIts, ClientCodeCoverage.
Still in progress: 
AbstractProject (about 66% complete), WhiteBoard, WhiteBoardColumn.

Notes: 
NoteCard is still having the server error problem (won't run because of communication failure. 
To add tests to ClientCodeCoverage, follow the convention seen in the file. If a test does not run on its own (like NoteCard) it will not work under ClientCodeCoverage. I still added those that do not run on their own anyways.
AbstractProject follows similarly  to NoteCard, in that it will not run because of the server communication error/problem.